Cloud-based LOcally linear Unbiased Dysbiosis (CLOUD) test

Cloud-based LOcally linear Unbiased Dysbiosis (CLOUD) test

Arguments

x

A phyloseq object

reference_samples

Vector of samples to use as reference.

dist_mat

A distance matrix. Can be output of phyloseq::distance or vegan::vegdist.

k_num

Neighbors to use. Default is 80 percent. User can define percent of samples. See Montassier E et al. 2018 for more details.

ndim

Dimension of the space in which the data are to be represented. Default is -1. See Montassier E et al. 2018

Value

A data frame with CLOUD stats and sample information

Details

Calculates CLOUD score.

Cloud-based LOcally linear Unbiased Dysbiosis (CLOUD) test is a non-parametric test and returns a measure of dysbiosis. The function was adapted from the original article by Montassier E et al. 2018. Here, a user defines a set of reference samples from which distance of every other sample is calculated. When calculating the CLOUD stats the k is an important parameter specified by argument k_num. By default we use a conservative 80 percent of the samples in each group.

References

Lloyd-Price J, Arze C, Ananthakrishnan AN et al. (2019). Multi-omics of the gut microbial ecosystem in inflammatory bowel diseases. Nature, 569(7758), pp.655-662.

Montassier E et al. (2018). CLOUD: a non-parametric detection test for microbiome outliers. Microbiome, 6(1), pp.1-10.

Author

Sudarshan A. Shetty

Examples

data("WirbelJ_2018")
library(phyloseq)
ps <- WirbelJ_2018
# Define controls as reference samples
ref.samples <- sample_names(subset_samples(WirbelJ_2018,
                                           disease == "healthy"))
dist.data <- phyloseq::distance(ps, "bray")
cloud.results <- cloudStatistic(x= ps,
                                dist_mat = dist.data,
                                reference_samples = ref.samples,
                                ndim=-1)
head(cloud.results)
#>                         stats    pvals                 ids   log2Stats
#> MMPU29365221ST      0.9803426 0.500000      MMPU29365221ST -0.02864206
#> MMPU68403337ST      0.9035595 0.890625      MMPU68403337ST -0.14630854
#> MMPU72854103ST      0.9411051 0.640625      MMPU72854103ST -0.08757229
#> MMPU84450604ST      0.8285715 1.000000      MMPU84450604ST -0.27130195
#> MMPU99077057ST      0.8640467 0.984375      MMPU99077057ST -0.21081880
#> CCMD10032470ST-11-0 0.9033809 0.906250 CCMD10032470ST-11-0 -0.14659362
#>                       study_name          subject_id body_site study_condition
#> MMPU29365221ST      WirbelJ_2018 CCMD11006829ST-21-0     stool             CRC
#> MMPU68403337ST      WirbelJ_2018 CCMD12232071ST-21-0     stool             CRC
#> MMPU72854103ST      WirbelJ_2018 CCMD13071240ST-21-0     stool             CRC
#> MMPU84450604ST      WirbelJ_2018 CCMD13934959ST-21-0     stool             CRC
#> MMPU99077057ST      WirbelJ_2018 CCMD14479708ST-21-0     stool             CRC
#> CCMD10032470ST-11-0 WirbelJ_2018 CCMD18872694ST-21-0     stool             CRC
#>                     disease age age_category gender country non_westernized
#> MMPU29365221ST          CRC  42        adult female     DEU              no
#> MMPU68403337ST          CRC  75       senior   male     DEU              no
#> MMPU72854103ST          CRC  66       senior female     DEU              no
#> MMPU84450604ST          CRC  56        adult   male     DEU              no
#> MMPU99077057ST          CRC  74       senior   male     DEU              no
#> CCMD10032470ST-11-0     CRC  63        adult   male     DEU              no
#>                     sequencing_platform DNA_extraction_kit     PMID
#> MMPU29365221ST            IlluminaHiSeq              Gnome 30936547
#> MMPU68403337ST            IlluminaHiSeq              Gnome 30936547
#> MMPU72854103ST            IlluminaHiSeq              Gnome 30936547
#> MMPU84450604ST            IlluminaHiSeq              Gnome 30936547
#> MMPU99077057ST            IlluminaHiSeq              Gnome 30936547
#> CCMD10032470ST-11-0       IlluminaHiSeq              Gnome 30936547
#>                     number_reads number_bases minimum_read_length
#> MMPU29365221ST          83456496   7310970279                  45
#> MMPU68403337ST          65269931   5370021950                   2
#> MMPU72854103ST          63427722   5486983593                  45
#> MMPU84450604ST          36979669   3186342244                  45
#> MMPU99077057ST          65475493   5637718108                  45
#> CCMD10032470ST-11-0     55300211   4780095276                  45
#>                     median_read_length NCBI_accession      curator BMI    tnm
#> MMPU29365221ST                      94           <NA> Paolo_Manghi  35 t3n0m0
#> MMPU68403337ST                      88           <NA> Paolo_Manghi  28 t3n2m0
#> MMPU72854103ST                      93           <NA> Paolo_Manghi  23 t3n0m0
#> MMPU84450604ST                      93           <NA> Paolo_Manghi  27 t3n0m0
#> MMPU99077057ST                      92           <NA> Paolo_Manghi  26 t3n0m1
#> CCMD10032470ST-11-0                 93           <NA> Paolo_Manghi  25 t3n1m0
#>                     ajcc nspecies group
#> MMPU29365221ST        ii      144  case
#> MMPU68403337ST       iii      110  case
#> MMPU72854103ST        ii      126  case
#> MMPU84450604ST        ii      100  case
#> MMPU99077057ST        iv      110  case
#> CCMD10032470ST-11-0  iii      145  case